Ai radar/requirements.txt
annotated-types==0.7.0
anthropic==0.84.0
anyio==4.12.1
blinker==1.9.0
certifi==2026.2.25
click==8.1.8
cloudpickle==3.1.2
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
docstring_parser==0.17.0
exceptiongroup==1.3.1
Flask==3.1.3
fonttools==4.60.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
kiwisolver==1.4.7
llvmlite==0.43.0
MarkupSafe==3.0.3
matplotlib==3.9.4
numba==0.60.0
numpy==2.0.2
packaging==26.0
pandas==2.3.3
pillow==11.3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.1.post1
scikit-learn==1.6.1
scipy==1.13.1
shap==0.49.1
six==1.17.0
slicer==0.0.8
sniffio==1.3.1
threadpoolctl==3.6.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
Werkzeug==3.1.6
zipp==3.23.0
